fix: non login user can not access profile page
