+++ packages/toolkit/src/components/top-bar/AppTopbar.tsx
@@ -3,8 +3,6 @@
 import cn from "clsx";
 import Link from "next/link";
 import React, { ReactElement } from "react";
-import { Icons } from "@instill-ai/design-system";
-import { TopbarLink } from "./TopbarLink";
 import {
   InstillStore,
   useAuthenticatedUser,
@@ -15,33 +13,7 @@ import { CloudTopbarDropdown } from "./CloudTopbarDropdown";
 import { CETopbarDropdown } from "./CETopbarDropdown";
 import { env } from "../../server";
 import { usePathname, useRouter } from "next/navigation";
-
-const topbarItems = [
-  {
-    pathName: "pipelines",
-    icon: <Icons.Pipeline className="h-6 w-6 stroke-semantic-fg-primary" />,
-    name: "Pipelines",
-  },
-  {
-    pathName: "connectors",
-    icon: (
-      <Icons.IntersectSquare className="h-6 w-6 stroke-semantic-fg-primary" />
-    ),
-    name: "Connectors",
-  },
-  {
-    pathName: "models",
-    icon: <Icons.Cube01 className="h-6 w-6 stroke-semantic-fg-primary" />,
-    name: "Models",
-  },
-  {
-    pathName: "dashboard",
-    icon: (
-      <Icons.BarChartSquare02 className="h-6 w-6 stroke-semantic-fg-primary" />
-    ),
-    name: "Dashboard",
-  },
-];
+import { TopbarLinks } from "./TopbarLinks";
 
 const selector = (store: InstillStore) => ({
   accessToken: store.accessToken,
@@ -60,7 +32,7 @@ export const AppTopbar = ({
   disabledUserDropdown?: boolean;
 }) => {
   const router = useRouter();
-  const pathName = usePathname();
+  const pathname = usePathname();
 
   const { accessToken, enabledQuery } = useInstillStore(useShallow(selector));
 
@@ -83,32 +55,9 @@ export const AppTopbar = ({
 
         {children ? (
           <div className="flex w-full flex-1 flex-row">{children}</div>
-        ) : (
-          <React.Fragment>
-            {env("NEXT_PUBLIC_APP_ENV") === "CLOUD" ? (
-              <TopbarLink
-                key="hub"
-                href="/hub"
-                icon={
-                  <Icons.CubeOutline className="h-6 w-6 stroke-semantic-fg-primary" />
-                }
-                name="Hub"
-                hightlighted={pathName.split("/")[1] === "hub"}
-                className="mx-1 my-2 px-4"
-              />
-            ) : null}
-            {topbarItems.map(({ pathName, name, icon }) => (
-              <TopbarLink
-                key={pathName}
-                href={`/${me.data?.id}/${pathName}`}
-                icon={icon}
-                name={name}
-                hightlighted={pathName.split("/")[2] === pathName}
-                className="mx-1 my-2 px-4"
-              />
-            ))}
-          </React.Fragment>
-        )}
+        ) : me.isSuccess ? (
+          <TopbarLinks entity={me.data.id} pathname={pathname} />
+        ) : null}
       </div>
 
       {disabledUserDropdown ? null : (
